% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.trtsel.R
\name{calibrate}
\alias{calibrate}
\title{method for calibrating trtsel objects}
\usage{
calibrate(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{not used}
}
\description{
method for calibrating trtsel objects
}
\seealso{
\code{\link{calibrate.trtsel}}
}
