% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.trtsel.R
\name{evaluate}
\alias{evaluate}
\title{a method for evaluating a trtsel object}
\usage{
evaluate(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{not used}
}
\description{
a method for evaluating a trtsel object
}
\seealso{
\code{\link{evaluate.trtsel}}
}
