\name{is.trtsel}
\alias{is.trtsel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{

is x a trtsel object?
}
\description{
 Function to check if an object is of class "trtsel"
}
\usage{
is.trtsel(x)
}

\arguments{
  \item{x}{
  Any R object. 
}
}

\value{
 Returns TRUE or FALSE depending on whether x is a "trtsel" object or not. 
}
