\name{print.calibrate.trtsel}

\alias{print.calibrate.trtsel}

\title{

print a trtsel object
}
\description{
  S3 method for class ''trtsel''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{calibrate.trtsel}(x, ...)
}

\arguments{
  \item{x}{
   object of class "calibrate.trtsel", output from the calibrate.trtsel function 
}
\item{\dots}{ ignored}
}
