% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trtsel.R
\name{print.compare.trtsel}
\alias{print.compare.trtsel}
\title{print a trtsel object}
\usage{
\method{print}{compare.trtsel}(x, ...)
}
\arguments{
\item{x}{object of class "compare.trtsel", output from the compare.trtsel
function}

\item{\dots}{ignored}
}
\description{
S3 method for class ''trtsel''
}
