\name{print.eval.trtsel}

\alias{print.eval.trtsel}

\title{

print a trtsel object
}
\description{
  S3 method for class ''trtsel''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{eval.trtsel}(x, ...)
}

\arguments{
  \item{x}{
   object of class "eval.trtsel", output from the eval.trtsel function 
}
\item{\dots}{ ignored}
}
