% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trtsel.R
\name{print.trtsel}
\alias{print.trtsel}
\title{print a trtsel object}
\usage{
\method{print}{trtsel}(x, ...)
}
\arguments{
\item{x}{object of class "trtsel", which can be created using function
"trtsel." See ?trtsel for more information.}

\item{\dots}{ignored}
}
\description{
S3 method for class ''trtsel''
}
