% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_measures.R
\name{print.trtsel_summary_measures}
\alias{print.trtsel_summary_measures}
\title{print output from 'trtsel_measures'}
\usage{
\method{print}{trtsel_summary_measures}(x, ...)
}
\arguments{
\item{x}{object of class "trtsel_summary_measures", output from the estimate_measures function}

\item{\dots}{ignored}
}
\description{
S3 method for class ''trtsel_measures''
}

