% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.trtsel.R
\name{is.trtsel}
\alias{is.trtsel}
\title{is x a trtsel object?}
\usage{
is.trtsel(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Returns TRUE or FALSE depending on whether x is a "trtsel" object or
not.
}
\description{
Function to check if an object is of class "trtsel"
}

