% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trtsel.R
\name{print.calibrate.trtsel}
\alias{print.calibrate.trtsel}
\title{print a trtsel object}
\usage{
\method{print}{calibrate.trtsel}(x, ...)
}
\arguments{
\item{x}{object of class "calibrate.trtsel", output from the
calibrate.trtsel function}

\item{\dots}{ignored}
}
\description{
S3 method for class ''trtsel''
}

