\name{print.compare.trtsel}

\alias{print.compare.trtsel}

\title{

print a trtsel object
}
\description{
  S3 method for class ''trtsel''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{compare.trtsel}(x, ...)
}

\arguments{
  \item{x}{
   object of class "compare.trtsel", output from the compare.trtsel function 
}
\item{\dots}{ ignored}
}
