% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trtsel.R
\name{print.eval.trtsel}
\alias{print.eval.trtsel}
\title{print a trtsel object}
\usage{
\method{print}{eval.trtsel}(x, ...)
}
\arguments{
\item{x}{object of class "eval.trtsel", output from the eval.trtsel function}

\item{\dots}{ignored}
}
\description{
S3 method for class ''trtsel''
}

