\name{print.trtsel}

\alias{print.trtsel}

\title{

print a trtsel object
}
\description{
  S3 method for class ''trtsel''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{trtsel}(x, ...)
}

\arguments{
  \item{x}{
   object of class "trtsel", which can be created using function "trtsel." See ?trtsel for more information. 
}
\item{\dots}{ ignored}
}
