\name{TreatmentSelection-package}
\alias{TreatmentSelection-package}
\alias{TreatmentSelection}
\docType{package}
\title{

Evaluate markers used to guide patient treatment selection decisions
}
\description{A suite of descriptive and inferential methods designed to evaluate
individual treatment selection markers and to compare candidate markers.  


}
\details{
\tabular{ll}{
Package: \tab TreatmentSelection\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2013-09-19\cr
License: \tab GPL-2\cr
}


}
\author{
Marshall Brown and Holly Janes

Maintainer: Marshall Brown <mdbrown@fhcrc.org>

}
\references{

Janes, Holly; Brown, Marshall D.; Pepe, Margaret; and Huang, Ying, "Statistical Methods for Evaluating and Comparing Biomarkers for Patient Treatment Selection" (January 2013). UW Biostatistics Working Paper Series. Working Paper 389.
http://biostats.bepress.com/uwbiostat/paper389
}

\seealso{
\code{\link{trtsel}} for creating trtsel objects, \code{\link{plot.trtsel}} for plotting risk curves and more, \code{\link{eval.trtsel}} for evaluating marker performance,  \code{\link{calibrate.trtsel}} for assessing model calibration, and \code{\link{compare.trtsel}} to compare two trtsel object. 
}



\keyword{ package }

\examples{
# See individual function man pages for examples. 

?trtsel
?plot.trtsel
?eval.trtsel
?compare.trtsel
?calibrate.trtsel
}
