% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentSelection-package.R
\docType{package}
\name{TreatmentSelection-package}
\alias{TreatmentSelection}
\alias{TreatmentSelection-package}
\title{Evaluate markers used to guide patient treatment selection decisions}
\description{
A suite of descriptive and inferential methods designed to evaluate
individual treatment selection markers and to compare candidate markers.
}
\details{
\tabular{ll}{ Package: \tab TreatmentSelection\cr Type: \tab Package\cr
Version: \tab 2.0.-\cr Date: \tab 2016-07-28\cr License: \tab GPL-2\cr }
}
\examples{

# See individual function man pages for examples. 

?trtsel
?plot.trtsel
?eval.trtsel
?compare.trtsel
?calibrate.trtsel

}
\author{
Marshall Brown and Holly Janes

Maintainer: Marshall Brown <mdbrown@fhcrc.org>
}
\references{
Janes, Holly; Brown, Marshall D; Pepe, Margaret; Huang, Ying; "An Approach
to Evaluating and Comparing Biomarkers for Patient Treatment Selection" The
International Journal of Biostatistics. Volume 0, Issue 0, ISSN (Online)
1557-4679, ISSN (Print) 2194-573X, DOI: 10.1515/ijb-2012-0052, April 2014
}
\seealso{
\code{\link{trtsel}} for creating trtsel objects,
\code{\link{plot.trtsel}} for plotting risk curves and more,
\code{\link{evaluate.trtsel}} for evaluating marker performance,
\code{\link{calibrate.trtsel}} for assessing model calibration, and
\code{\link{compare.trtsel}} to compare two trtsel object.
}
\keyword{package}

